+++ src/db/index/common/doc.cc
@@ -187,45 +187,6 @@ template <class... Ts>
 overloaded(Ts...) -> overloaded<Ts...>;
 
 
-bool sort_and_find_duplicates(uint32_t *indices, char *values, size_t n,
-                              size_t value_byte_size) {
-  if (n <= 1) {
-    return false;
-  }
-  bool already_sorted = true;
-  for (size_t i = 1; i < n; ++i) {
-    if (indices[i] == indices[i - 1]) {
-      return true;
-    }
-    if (indices[i] < indices[i - 1]) {
-      already_sorted = false;
-      break;
-    }
-  }
-  if (already_sorted) {
-    return false;
-  }
-  std::vector<size_t> perm(n);
-  std::iota(perm.begin(), perm.end(), size_t{0});
-  std::sort(perm.begin(), perm.end(),
-            [&](size_t a, size_t b) { return indices[a] < indices[b]; });
-  std::vector<uint32_t> sorted_indices(n);
-  std::vector<char> sorted_values(n * value_byte_size);
-  for (size_t i = 0; i < n; ++i) {
-    sorted_indices[i] = indices[perm[i]];
-    std::memcpy(sorted_values.data() + i * value_byte_size,
-                values + perm[i] * value_byte_size, value_byte_size);
-  }
-  std::memcpy(indices, sorted_indices.data(), n * sizeof(uint32_t));
-  std::memcpy(values, sorted_values.data(), n * value_byte_size);
-  for (size_t i = 1; i < n; ++i) {
-    if (indices[i] == indices[i - 1]) {
-      return true;
-    }
-  }
-  return false;
-}
-
 }  // namespace
 
 
@@ -1267,146 +1228,6 @@ bool Doc::operator==(const Doc &other) const {
   }
 
   return true;
-}
-
-Status SearchQuery::validate_and_sanitize(const FieldSchema *schema) {
-  if ((uint32_t)topk_ > kMaxQueryTopk) {
-    return Status::InvalidArgument("Invalid query: topk[", topk_,
-                                   "] exceeds the maximum allowed value of ",
-                                   kMaxQueryTopk);
-  }
-  if (output_fields_.has_value() &&
-      output_fields_->size() > kMaxOutputFieldSize) {
-    return Status::InvalidArgument(
-        "Invalid query: too many output fields, the maximum allowed is ",
-        kMaxOutputFieldSize);
-  }
-
-  auto *vc = target_.get_vector_clause();
-  auto &field_name_ = target_.field_name_;
-  // A "scalar-only filter" query has no vector payload — either the clause
-  // is not a VectorClause (e.g., FtsClause) or its fields are all empty.
-  bool no_vector_payload = (vc == nullptr) || (vc->query_vector_.empty() &&
-                                               vc->sparse_indices_.empty());
-
-  if (schema == nullptr) {
-    if (no_vector_payload) {
-      // Scalar-only filter query
-      return Status::OK();
-    } else {
-      // If a query vector was provided, the field must exist as a vector field
-      // since we are performing a vector similarity search.
-      return Status::InvalidArgument(
-          "Invalid query: query vector is provided, but query field[",
-          field_name_,
-          "] does not exist or is not a vector field in the collection");
-    }
-  }
-
-  // Schema is non-null from here on: a vector payload is required.
-  if (no_vector_payload) {
-    return Status::InvalidArgument("Invalid query: missing query clause for [",
-                                   field_name_, "]");
-  }
-
-  auto &query_vector_ = vc->query_vector_;
-  auto &query_sparse_indices_ = vc->sparse_indices_;
-  auto &query_sparse_values_ = vc->sparse_values_;
-  auto &query_params_ = target_.query_params_;
-
-  // Vector query
-  if (schema->is_dense_vector()) {
-    // Validate dimension
-    auto dim = schema->dimension();
-    switch (schema->data_type()) {
-      case DataType::VECTOR_FP16:
-        if (dim * sizeof(float16_t) != query_vector_.size()) {
-          return Status::InvalidArgument(
-              "Invalid query: dimension mismatch, expected ", dim, " but got ",
-              query_vector_.size() / sizeof(float16_t), " (FP16)");
-        }
-        break;
-      case DataType::VECTOR_FP32:
-        if (dim * sizeof(float) != query_vector_.size()) {
-          return Status::InvalidArgument(
-              "Invalid query: dimension mismatch, expected ", dim, " but got ",
-              query_vector_.size() / sizeof(float), " (FP32)");
-        }
-        break;
-      case DataType::VECTOR_FP64:
-        if (dim * sizeof(double) != query_vector_.size()) {
-          return Status::InvalidArgument(
-              "Invalid query: dimension mismatch, expected ", dim, " but got ",
-              query_vector_.size() / sizeof(double), " (FP64)");
-        }
-        break;
-      case DataType::VECTOR_INT8:
-        if (dim * sizeof(int8_t) != query_vector_.size()) {
-          return Status::InvalidArgument(
-              "Invalid query: dimension mismatch, expected ", dim, " but got ",
-              query_vector_.size() / sizeof(int8_t), " (INT8)");
-        }
-        break;
-      case DataType::VECTOR_INT16:
-      case DataType::VECTOR_INT4:
-      case DataType::VECTOR_BINARY32:
-      case DataType::VECTOR_BINARY64:
-        return Status::NotSupported(
-            "Invalid query: dense vector type of field[", field_name_,
-            "] is not supported");
-      default:
-        return Status::InvalidArgument("Invalid query: field[", field_name_,
-                                       "] is not a dense vector field");
-    }
-  } else if (schema->is_sparse_vector()) {
-    size_t value_byte_size = 0;
-    switch (schema->data_type()) {
-      case DataType::SPARSE_VECTOR_FP32:
-        value_byte_size = sizeof(float);
-        break;
-      case DataType::SPARSE_VECTOR_FP16:
-        value_byte_size = sizeof(float16_t);
-        break;
-      default:
-        return Status::InvalidArgument(
-            "Invalid query: sparse vector type of field[", field_name_,
-            "] is not supported");
-    }
-    if (query_sparse_indices_.size() % sizeof(uint32_t) != 0 ||
-        query_sparse_values_.size() % value_byte_size != 0 ||
-        query_sparse_indices_.size() / sizeof(uint32_t) !=
-            query_sparse_values_.size() / value_byte_size) {
-      return Status::InvalidArgument(
-          "Invalid query: sparse vector query for field[", field_name_,
-          "] has mismatched indices and values sizes");
-    }
-    size_t n_indices = query_sparse_indices_.size() / sizeof(uint32_t);
-    if (n_indices > kSparseMaxDimSize) {
-      return Status::InvalidArgument(
-          "Invalid query: too many sparse indices, the maximum allowed is ",
-          kSparseMaxDimSize);
-    }
-    if (sort_and_find_duplicates(
-            reinterpret_cast<uint32_t *>(query_sparse_indices_.data()),
-            query_sparse_values_.data(), n_indices, value_byte_size)) {
-      return Status::InvalidArgument(
-          "Invalid query: sparse vector query for field[", field_name_,
-          "] contains duplicate indices");
-    }
-  } else {
-    return Status::InvalidArgument("Invalid query: field[", field_name_,
-                                   "] is not a vector field");
-  }
-  // Validate query_params type
-  if (query_params_ && query_params_->type() != schema->index_type()) {
-    return Status::InvalidArgument(
-        "Invalid query: query params type does not match the index type of "
-        "vector field[",
-        field_name_, "], expected ",
-        IndexTypeCodeBook::AsString(schema->index_type()), " but got ",
-        IndexTypeCodeBook::AsString(query_params_->type()));
-  }
-  return Status::OK();
 }
 
 }  // namespace zvec
